+++ app/inscriptions/create/page.tsx
@@ -1,8 +1,13 @@
 import { NextPage } from 'next';
 import { InscriptionsCreate } from '../components/inscription-create';
+import { Suspense } from 'react';
 
 const InscriptionsCreatePage: NextPage = () => {
-  return <InscriptionsCreate />;
+  return (
+    <Suspense>
+      <InscriptionsCreate />
+    </Suspense>
+  );
 };
 
 export default InscriptionsCreatePage;