+++ components/header-chain-indicator.tsx
@@ -6,13 +6,14 @@ export const HeaderChainIndicator = () => {
   const { chainType } = useConfig();
 
   return (
-    <div className="flex items-center justify-center gap-1 text-xs bg-slate-100 dark:bg-slate-900 py-1 text-center font-light">
-      <div className="flex h-2 w-2 mt-[-1px] rounded-full bg-green-700 dark:bg-green-200 mr-1" />{' '}
-      <div>MultiversX</div>{' '}
-      <div className="capitalize font-bold">{chainType}</div>
-      <div>|</div>
+    <div className="flex items-center flex-col sm:flex-row justify-center gap-1 text-xs bg-zinc-100 dark:bg-zinc-900 py-1 text-center font-light">
+      <div className="flex items-center justify-center gap-1">
+        <div className="flex h-2 w-2 mt-[-1px] rounded-full bg-green-700 dark:bg-green-200 mr-1" />{' '}
+        <div>You are using MultiversX</div>{' '}
+        <div className="capitalize font-bold">{chainType}</div>
+      </div>
       <div>
-        Switch to{' '}
+        (Do you want to switch to{' '}
         <a
           href={
             chainType === 'mainnet'
@@ -24,6 +25,7 @@ export const HeaderChainIndicator = () => {
         >
           {chainType === 'mainnet' ? 'devnet' : 'mainnet'}
         </a>
+        ?)
       </div>
     </div>
   );