+++ components/operations/non-fungible-tokens/send.tsx
@@ -1,6 +1,5 @@
 import * as z from 'zod';
 import { useForm } from 'react-hook-form';
-import axios from 'axios';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Form } from '@/components/ui/form';
 import {
@@ -14,7 +13,7 @@ import { OperationsSubmitButton } from '@/components/operations/operations-submi
 import { useContext } from 'react';
 import { OperationsStateDialogContext } from '@/components/operations/operations-status-dialog';
 import { OperationContentProps } from '@/components/operations/operations-common-types';
-import { ScTokenTransferType, useConfig } from '@useelven/core';
+import { ESDTType } from '@useelven/core';
 import { transfersOperationsGasLimit } from '@/components/operations/constants';
 
 const formSchema = z.object({
@@ -26,7 +25,6 @@ export const Send = ({ transfer, close }: OperationContentProps) => {
   const { setOpen: setTxStatusDialogOpen } = useContext(
     OperationsStateDialogContext
   );
-  const { apiAddress } = useConfig();
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
@@ -37,47 +35,16 @@ export const Send = ({ transfer, close }: OperationContentProps) => {
   });
 
   const onSubmit = async ({ tokenId, address }: z.infer<typeof formSchema>) => {
-    try {
-      // TODO: replace with useElven useApiCall when ready to handle such cases
-      const nftOnNetwork = await axios.get<{
-        nonce: number;
-        collection: string;
-      }>(`${apiAddress}/nfts/${tokenId.trim()}`, {
-        headers: {
-          'Content-Type': 'application/json',
-          Accept: 'application/json',
-        },
-      });
+    transfer?.({
+      type: ESDTType.NonFungibleESDT,
+      tokenId,
+      receiver: address.trim(),
+      gasLimit: transfersOperationsGasLimit,
+    });
 
-      const nonce = nftOnNetwork?.data?.nonce;
-      const collectionId = nftOnNetwork?.data?.collection;
-
-      // TODO: show the error in the transaction status modal
-      if (!nonce || !collectionId) {
-        console.error(
-          "Can't read the nonce or/and collection id of the token, using MultiversX API!"
-        );
-        return;
-      }
-
-      transfer?.({
-        type: ScTokenTransferType.ESDTNFTTransfer,
-        tokenId: collectionId,
-        address: address.trim(),
-        amount: '1',
-        gasLimit: transfersOperationsGasLimit,
-        nonce,
-      });
-
-      setTxStatusDialogOpen(true);
-      form.reset();
-      close();
-    } catch (e) {
-      console.error(
-        "Can't read the nonce or/and collection id of the token, using MultiversX API!",
-        e
-      );
-    }
+    setTxStatusDialogOpen(true);
+    form.reset();
+    close();
   };
 
   return (