+++ components/operations/operations-content-map.tsx
@@ -66,11 +66,9 @@ type OperationsContentMapProps = {
   transfer?: ({
     type,
     tokenId,
-    nonce,
     gasLimit,
-    address,
+    receiver,
     amount,
-    value,
     endpointName,
     endpointArgs,
   }: ScTokenTransferArgs) => void;