+++ lib/get-token-id.ts
@@ -1,0 +1,5 @@
+import { CommonOpertationContentProps } from '@/components/operations/operations-common-types';
+
+export const getTokenIdKey = (
+  tokenType: CommonOpertationContentProps['tokenType']
+) => (tokenType === 'fungible' ? 'identifier' : 'collection');